refac: formatDate util
