+++ src/app/team/records/[meetingId]/page.tsx
@@ -19,13 +19,13 @@ import {
   PopoverClose,
 } from '@/components/internal/ui/popover';
 import { toast } from 'sonner';
-import ConfirmModal from '@/app/calendar/[id]/ConfirmModal';
 import EnhancedAudioPlayer, { AudioPlayerHandle } from '@/components/EnhancedAudioPlayer';
 import ScriptTranscript from '@/components/ScriptTranscript';
 import { useMeetingSummary, useMeetingRecommendations } from '@/hooks/useMeeting';
 import { useBookmark } from '@/hooks/useBookmark';
 import RecommandSection from '@/components/RecommandSection';
 import { useActiveTeam } from '@/context/activeTeamContext';
+import { formatKoreanDate } from '@/utils/fotmatDate';
 
 interface AgendaItem {
   agenda: string;
@@ -413,17 +413,6 @@ export default function MeetingDetailPage() {
       </div>
     );
   }
-
-  const formatKoreanDate = (isoString: string): string => {
-    const date = new Date(isoString);
-    const year = date.getFullYear();
-    const month = (date.getMonth() + 1).toString().padStart(2, '0');
-    const day = date.getDate().toString().padStart(2, '0');
-    const dayOfWeek = ['일', '월', '화', '수', '목', '금', '토'][date.getDay()];
-    const hours = date.getHours().toString().padStart(2, '0');
-    const minutes = date.getMinutes().toString().padStart(2, '0');
-    return `${year}년 ${month}월 ${day}일 (${dayOfWeek}) ${hours}:${minutes}`;
-  };
 
   console.log('EnhancedAudioPlayer에 전달될 데이터 확인:', meetingDetail);
