+++ src/utils/fotmatDate.ts
@@ -1,0 +1,12 @@
+export const formatKoreanDate = (isoString: string): string => {
+  if (!isoString) return '';
+  const date = new Date(isoString);
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, '0');
+  const day = date.getDate().toString().padStart(2, '0');
+  const dayOfWeek = ['일', '월', '화', '수', '목', '금', '토'][date.getDay()];
+  const hours = date.getHours().toString().padStart(2, '0');
+  const minutes = date.getMinutes().toString().padStart(2, '0');
+
+  return `${year}년 ${month}월 ${day}일 (${dayOfWeek}) ${hours}:${minutes}`;
+};
